hash password

+++ controllers/news.controllers.js
@@ -15,6 +15,12 @@ export const news = {
 
 };
 
+/**
+ * display form for create news
+ * @param req
+ * @param res
+ * @param next
+ */
 function createDisplay(req, res, next) {
   console.log(`User authenticated? ${req.isAuthenticated()}`);
   if (isAuthorised(req)) {
@@ -24,6 +30,12 @@ function createDisplay(req, res, next) {
   }
 }
 
+/**
+ * diplay form for update news
+ * @param req
+ * @param res
+ * @param next
+ */
 function edit(req, res, next) {
   console.log(`User authenticated? ${req.isAuthenticated()}`);
   if (isAuthorised(req)) {
@@ -33,6 +45,12 @@ function edit(req, res, next) {
   }
 }
 
+/**
+ * display page with one news
+ * @param req
+ * @param res
+ * @param next
+ */
 function displayOneNews(req, res, next) {
   NewsSchema.findById(req.params.id, function (err, response) {
     if (err) res.redirect('/news');
@@ -43,21 +61,42 @@ function displayOneNews(req, res, next) {
   })
 }
 
+/**
+ * verify if file is a image
+ * @param file
+ * @returns {boolean}
+ */
 function isFileImage(file) {
   if (!file) return false;
   const acceptedImageTypes = ['image/gif', 'image/jpeg', 'image/png'];
   return acceptedImageTypes.includes(file['mimetype']);
 }
 
+/**
+ * display all
+ * @param req
+ * @param res
+ * @param next
+ */
 function displayAll(req, res, next) {
   NewsSchema.find({}, [], {$orderby: {'created_at': -1}}, (err, post) => {
     res.render('news', {title: "blog", news: post, admin: isAuthorised(req), name: getName(req)})
   })
 }
 
+/**
+ * create news if user is admin
+ * @param req
+ * @param res
+ * @param next
+ * @returns {*}
+ */
 function create(req, res, next) {
   if (!isAuthorised(req)) {
     return res.json({response: "you need to be a admin"}).status(403)
+  }
+  if (!req.file){
+    res.json({error: "please put an image"})
   }
   const tempPath = req.file.path;
   const targetPath = tempPath.replace('/\\/g', '/').replace('public', '');
@@ -84,8 +123,7 @@ function create(req, res, next) {
       }
       res.status(200).json('NewsSchema Created successfully')
     })
-  } else {
-    console.log("test32");
+  } else if (req.file) {
     fs.unlink(tempPath, err => {
       return res
         .status(403)
@@ -93,8 +131,18 @@ function create(req, res, next) {
         .json({"error": "Only image files are allowed!"});
     });
   }
+  return res
+    .status(403)
+    .contentType("text/json")
+    .json({"error": "Only image files are allowed!"});
 }
 
+/**
+ * get a news
+ * @param req
+ * @param res
+ * @param next
+ */
 function get(req, res, next) {
   console.log(req.params.id);
   NewsSchema.findById(req.params.id, function (err, response) {
@@ -103,28 +151,47 @@ function get(req, res, next) {
   })
 }
 
+/**
+ * update news if user is admin
+ * @param req
+ * @param res
+ * @param next
+ * @returns {*}
+ * @private
+ */
 function _update(req, res, next) {
   if (!isAuthorised(req)) {
     return res.json({response: "you need to be a admin"}).status(403)
   }
-  const tempPath = req.file.path;
-  const targetPath = tempPath.replace('/\\/g', '/').replace('public', '');
+  if (req.file) {
+    const tempPath = req.file.path;
+    req.body.image = tempPath.replace('/\\/g', '/').replace('public', '');
+  }
   if (isFileImage(req.file)) {
-    req.body.image = targetPath;
-    NewsSchema.findByIdAndUpdate(req.params.id, {$set: req.body}, function (err, response) {
-      if (err) return next(err);
-      res.json(response);
-    });
-  } else {
-    fs.unlink(tempPath, err => {
+    // req.body.image = targetPath;
+
+  } else if (req.file) {
+      fs.unlink(tempPath, err => {
       return res
         .status(403)
         .contentType("text/json")
         .json({"error": "Only image files are allowed!"});
     });
   }
+  NewsSchema.findByIdAndUpdate(req.params.id, {$set: req.body}, function (err, response) {
+    if (err) return next(err);
+    return res.json(response);
+  });
 }
 
+/**
+ * delete news if user is admin
+ * @param req
+ * @param res
+ * @param next
+ * @returns {*}
+ * @private
+ */
 function _delete(req, res, next) {
   if (!isAuthorised(req)) {
     return res.json({response: "you need to be a admin"}).status(403)
